Using useState hook at first time
We call it inside a function component to add some local state to it.
React will preserve this state between re-renders.
useState returns a pair: the current state value and a function that
lets you update it. You can call this function from an event handler or somewhere else. It’s similar to this.setState in a class, except it doesn’t merge the old and new state together.